// backend/src/common/enums/permission.enum.ts
export enum Permission {
  MANAGE_TENANTS = 'MANAGE_TENANTS',
  VIEW_PLATFORM_ANALYTICS = 'VIEW_PLATFORM_ANALYTICS',
  MANAGE_PLATFORM_BILLING = 'MANAGE_PLATFORM_BILLING',
  MANAGE_PLATFORM_THEMES = 'MANAGE_PLATFORM_THEMES',

  MANAGE_TENANT_USERS = 'MANAGE_TENANT_USERS',
  MANAGE_TENANT_WEBSITE = 'MANAGE_TENANT_WEBSITE',
  MANAGE_TENANT_PRODUCTS = 'MANAGE_TENANT_PRODUCTS',
  MANAGE_TENANT_ORDERS = 'MANAGE_TENANT_ORDERS',
  MANAGE_TENANT_CUSTOMERS = 'MANAGE_TENANT_CUSTOMERS',
  MANAGE_TENANT_BILLING = 'MANAGE_TENANT_BILLING',

  CRM_READ = 'CRM_READ',
  CRM_WRITE = 'CRM_WRITE',

  MANAGE_THEMES_GLOBAL = 'MANAGE_THEMES_GLOBAL',
  SELECT_TENANT_THEME = 'SELECT_TENANT_THEME',

  VIEW_TENANT_ANALYTICS = 'VIEW_TENANT_ANALYTICS',
}

export type PermissionType = keyof typeof Permission | Permission;